+++ src/main/java/org/infernus/idea/checkstyle/actions/GroupByConfigurationLocation.java
@@ -1,0 +1,10 @@
+package org.infernus.idea.checkstyle.actions;
+
+import org.infernus.idea.checkstyle.toolwindow.ResultGrouping;
+
+public class GroupByConfigurationLocation extends GroupingAction {
+
+    public GroupByConfigurationLocation() {
+        super(ResultGrouping.BY_CONFIGURATION_LOCATION);
+    }
+}